+++ FINLAY_SMITH/Asteroid.java
@@ -1,4 +1,4 @@
-package FINLAY_SMITH.GameAssets;
+package FINLAY_SMITH;
 
 public class Asteroid extends Body {
