+++ FINLAY_SMITH/Body.java
@@ -1,4 +1,4 @@
-package FINLAY_SMITH.GameAssets;
+package FINLAY_SMITH;
 
 public class Body {
     protected Vector position;
