+++ FINLAY_SMITH/GameArena.java
@@ -1,4 +1,4 @@
-package FINLAY_SMITH.GameAssets;
+package FINLAY_SMITH;
 
 import javax.swing.*;
 import java.awt.*;
