+++ FINLAY_SMITH/ScoreCounter.java
@@ -1,4 +1,4 @@
-package FINLAY_SMITH.GameAssets;
+package FINLAY_SMITH;
 
 public class ScoreCounter extends Text {
     private int scoreValue;
