+++ FINLAY_SMITH/Vector.java
@@ -1,4 +1,4 @@
-package FINLAY_SMITH.GameAssets;
+package FINLAY_SMITH;
 
 public class Vector {
     public double x;
